Update README.md
1. Write a Java program where the current year and the year in which an
employee joined the organization are entered through the keyboard. If the
number of years for which the employee has served the organization is
greater than 5, a bonus of Rs. 5000/- is given to the employee. If the years
of service are between 3 and 5 (inclusive), a bonus of Rs. 3000/- is given.
If the years of service are less than 3, then the program should print a
message indicating that no bonus is awarded.
2. A library charges a fine for every book returned late. For the first 7 days,
the fine is 50 paise, for 8-14 days the fine is one rupee, and above 14 days
fine is 5 rupees. If you return the book after 21 days, your membership will
be canceled. Write a program to accept the number of days the member is
late to return the book and display the fine or the appropriate message.
3. While purchasing certain items, a discount of 10% is offered if the quantity
purchased is more than 50. If the quantity purchased is between 25 and 50
(inclusive), a discount of 5% is offered. If the quantity purchased is less
than 25, no discount is offered. Write a program to calculate the total
expenses given the quantity and price per item entered through the
keyboard.
4. Write a Java program to check if a vowel is present in a string. Additionally,
the program should count the total number of vowels present and display
the count along with whether any vowels are present or not.
5. Write a Java program to check whether a given number is divisible by 7
and 13 simultaneously. Additionally, if the number is divisible by both 7 and
13, the program should display the quotient and remainder when the
number is divided by 7 and 13.
6. Write a Java program to enter numbers until the user wants, and at the
end, the program should display the largest and smallest numbers entered.
Additionally, the program should calculate and display the average of all
the numbers entered